refactor(advisor): split engine into advisor-engine.js, slim orchestrator
Extract recommend() and verdict() to advisor-engine.js as named exports,
drop the IIFE (ES module scope replaces it), make buildNav() return a
string consistent with the other builders, add buildVerdictHTML() to
slim composeComparison(), split setup() into setupQuestion/Result/Oos
helpers, and replace the hardcoded appType/mobile-desktop OOS escape
with a data-driven selectedOpt?.escape check.
